UPDATE/Repo_Service

+++ KoiFarmShop/KoiFarmShop.Repository/IRepo/IKoiFishRepository.cs
@@ -5,11 +5,9 @@ namespace KoiFarmShop.Repository;
 public interface IKoiFishRepository
 {
     Task<IEnumerable<KoiFish>> GetAllAsync();
-    Task<KoiFish> GetByIdAsync(int id);
+    Task<KoiFish> GetByIdAsync(long id);
     Task AddAsync(KoiFish koiFish);
     Task UpdateAsync(KoiFish koiFish);
     Task DeleteAsync(int id);
-
     long GetnextKoiFishId();
-
 }
